pyproject.toml
[project]
name = "text-to-image-fast"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pillow>=10.4",
    "python-dotenv>=1.0",
]

